# my-page-api/src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: JPro My Page API
  description: API for JPro My Page
  version: 1.0.0
paths:
...
  /me:
    get:
      tags:
        - User
      operationId: getMe
      summary: Get current user
      description: Returns the currently authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
...
  /booking:
    get:
      tags:
        - Booking
      operationId: getBookings
      summary: Get bookings
      description: Returns bookings within a date range
      parameters:
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
...
  /informationNotice:
    get:
      tags:
        - InformationNotice
      operationId: getInformationNotices
      summary: Get all information notices in a given period
      parameters:
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of information notices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InformationNotice'
        '400':
          description: Invalid date format
    post:
      tags:
        - InformationNotice
      operationId: createInformationNotice
      summary: Create a new information notice (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInformationNotice'
      responses:
        '201':
          description: Information notice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InformationNotice'
        '400':
          description: Bad request
  /informationNotice/{infoNoticeId}:
    patch:
      tags:
        - InformationNotice
      operationId: updateInformationNotice
      summary: Update an existing information notice (admin only)
      parameters:
        - name: infoNoticeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInformationNotice'
      responses:
        '200':
          description: Information notice updated
        '400':
          description: Bad request
        '404':
          description: Information notice not found
    delete:
      tags:
        - InformationNotice
      operationId: deleteInformationNotice
      summary: Delete an information notice (admin only)
      parameters:
        - name: infoNoticeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Information notice deleted
        '404':
          description: Information notice not found
  /informationNotice/vacancy:
    get:
      tags:
        - InformationNotice
      operationId: getInformationNoticeVacancies
      summary: Get information notice vacancies in a time period
      parameters:
        - in: query
          name: startdate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: enddate
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of vacant dates
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: date
        '400':
          description: Invalid date format
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        givenName:
          type: string
        familyName:
          type: string
        sub:
          type: string
        admin:
          type: boolean
        enabled:
          type: boolean
        employeeNumber:
          type: integer
      required:
        - id
        - email
...
    Booking:
      type: object
      properties:
        id:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        apartment:
          $ref: '#/components/schemas/Apartment'
        employeeName:
          type: string
        isPending:
          type: boolean
      required:
        - id
        - startDate
        - endDate
        - apartment
    Apartment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cabin_name:
          type: string
        sort_order:
          type: integer
      required:
        - id
        - cabin_name
...
    InformationNotice:
      type: object
      properties:
        id:
          type: integer
          format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
      required:
        - startDate
        - endDate
        - description
    CreateInformationNotice:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
      required:
        - startDate
        - endDate
        - description
    UpdateInformationNotice:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
      required:
        - startDate
        - endDate
        - description
...